Refactored product request validator
